please the picky compilers by staying with short as the data we get is short only

+++ ares/ares_query.c
@@ -92,15 +92,15 @@ static struct query* find_query_by_id(ares_channel channel, int id)
    performed per id generation. In practice this search should happen only
    once per newly generated id
 */
-static int generate_unique_id(ares_channel channel)
+static unsigned short generate_unique_id(ares_channel channel)
 {
-  int id;
+  short id;
 
   do {
 	id = ares__generate_new_id(&channel->id_key);
   } while (find_query_by_id(channel,id));
 
-  return id;
+  return (unsigned short)id;
 }
 
 void ares_query(ares_channel channel, const char *name, int dnsclass,
